refactor (Reports): add exceptions to EmployeeService

+++ Reports.Infrastructure/Services/EmployeeService.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using System.Collections.Generic;
 using Reports.Core.Entities;
+using Reports.Core.Tools;
 using Reports.Infrastructure.DAL;
 
 namespace Reports.Infrastructure.Services
@@ -15,6 +16,11 @@ namespace Reports.Infrastructure.Services
         }
         public Employee Create(string name, Guid? bossId = null)
         {
+            if (bossId != null && !_context.Employees.Any(employee => employee.Id == bossId))
+            {
+                throw new ReportsException($"Error: employee with id {bossId} doesn't exist");
+            }
+
             var newEmployee = new Employee(name, bossId);
             _context.Employees.Add(newEmployee);
             _context.SaveChanges();
@@ -23,6 +29,11 @@ namespace Reports.Infrastructure.Services
 
         public void Delete(Guid employeeId)
         {
+            if (!_context.Employees.Any(employee => employee.Id == employeeId))
+            {
+                throw new ReportsException($"Error: employee with id {employeeId} doesn't exist");
+            }
+
             Employee employee = GetById(employeeId);
             _context.Employees.Remove(employee);
             _context.SaveChanges();
@@ -35,11 +46,27 @@ namespace Reports.Infrastructure.Services
 
         public Employee GetById(Guid id)
         {
+            if (!_context.Employees.Any(employee => employee.Id == id))
+            {
+                throw new ReportsException($"Error: employee with id {id} doesn't exist");
+            }
+
             return _context.Employees.FirstOrDefault(employee => employee.Id == id);
         }
 
         public Employee Update(Guid employeeId, Guid newBossId)
         {
+            if (!_context.Employees.Any(employee => employee.Id == employeeId))
+            {
+                throw new ReportsException($"Error: employee with id {employeeId} doesn't exist");
+            }
+
+            if (!_context.Employees.Any(employee => employee.Id == newBossId))
+            {
+                throw new ReportsException($"Error: employee with id {newBossId} doesn't exist");
+            }
+
+
             Employee employee = GetById(employeeId);
             employee.UpdateBoss(newBossId);
             _context.SaveChanges();
